+++ src/system/Device.js
@@ -1,20 +1,15 @@
-import { Black } from "../Black";
-
-/**
- * @type {number|null}
- */
-let mPixelRatioCache = null;
-
 /**
  * Contains system functions.
  * @static
  * @cat system
  */
 export class Device {
-  /**
-   * Static class.
-   */
   constructor() {
+    /**
+     * @type {number|null}
+     * @private
+     */
+    this.mPixelRatioCache = null;
   }
 
   /**
@@ -22,7 +17,7 @@ export class Device {
    * 
    * @return {string}
    */
-  static get os() {
+  get os() {
     let userAgent = navigator.userAgent || navigator.vendor || window.opera;
 
     if (/windows phone/i.test(userAgent))
@@ -42,7 +37,7 @@ export class Device {
    *
    * @return {boolean}
    */
-  static get isTouch() {
+  get isTouch() {
     let hasEvent = 'ontouchstart' in window;
     if (hasEvent)
       return true;
@@ -58,7 +53,7 @@ export class Device {
    *
    * @return {boolean}
    */
-  static get isMobile() {
+  get isMobile() {
     return /Mobi/.test(navigator.userAgent);
   }
 
@@ -67,11 +62,11 @@ export class Device {
    *
    * @return {number}
    */
-  static get pixelRatio() {
-    if (mPixelRatioCache === null)
-      mPixelRatioCache = Device.getDevicePixelRatio();
+  get pixelRatio() {
+    if (this.mPixelRatioCache === null)
+      this.mPixelRatioCache = this.getDevicePixelRatio();
 
-    return mPixelRatioCache;
+    return this.mPixelRatioCache;
   }
 
   /**
@@ -79,46 +74,22 @@ export class Device {
    *
    * @return {boolean}
    */
-  static get webAudioSupported() {
+  get webAudioSupported() {
     return window['AudioContext'] != null || window['webkitAudioContext'] != null;
   }
 
   /**
    * Returns device pixel ratio.
    *
-   * @static
    * @suppress {missingProperties}
    * @return {number} Description
    */
-  static getDevicePixelRatio() {
+  getDevicePixelRatio() {
     if (window.screen.systemXDPI !== undefined && window.screen.logicalXDPI !== undefined && window.screen.systemXDPI > window.screen.logicalXDPI)
       return window.screen.systemXDPI / window.screen.logicalXDPI;
     else if (window.devicePixelRatio !== undefined)
       return window.devicePixelRatio;
 
     return 1;
-  }
-
-  /**
-   * Returns true if device is in landscape orientation.
-   * 
-   * @public
-   * @readonly
-   * @returns {boolean}
-   */
-  static get isLandscape() {
-    let size = Black.engine.viewport.size;
-    return size.width >= size.height;
-  }
-
-  /**
-   * Returns true if device is in portrait orientation.
-   * 
-   * @public
-   * @readonly
-   * @returns {boolean}
-   */
-  static get isPortrait() {
-    return !Device.isLandscape;
   }
 }