+++ CHANGELOG.md
@@ -11,6 +11,8 @@ Black Engine: Changelog
 - Changed `Black` to `Black.engine
 - Changed bundler to Rollup
 - Changed Engine, MasterAudio, Input, Viewport to use `__initialize` method for initialization so constructor can be overriden
+- Changed `Device` to `Black.device`
+- Changed `Device#isLandscape` to `Black.viewport.isLandscape`, same for `isPortrait`.
 
 ### Bug Fixes
 - Fixed all circular dependencies